### 2023-03-07 #5226 @editorialbot
[REVIEW]: osiris: An R package to process climate impacts on agricultural yields for the Global Change Analysis Model
**Submitting author:** @hahsan1 (<a href="http://orcid.org/0000-0003-1631-1789">Hamza Ahsan</a>)
**Repository:** https://github.com/JGCRI/osiris
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.3
