### 2023-03-23 #5226 @jsun
Hi, @hahsan1

I tried to install the package and run `get_example_data()` and then ran the example script of Step 1. But I got the following errors. Would you please tell me where can I download the `tas_Amon_CanESM5_historical_r1i1p1f1_orig.nc` file? Thanks.

```
Starting calculate_deltas_from_climate...
Error in R_nc4_open: No such file or directory
Error in ncdf4::nc_open(batch0[1]) :
  Error in nc_open trying to open file Osiris_Data/climate_data/tas_Amon_CanESM5_historical_r1i1p1f1_orig.nc (return_on_error= FALSE )
In addition: Warning message:
'memory.limit()' is Windows-specific
```
