### 2023-03-23 #5226 @hahsan1
Hi @jsun,
Just to clarify, you are on devTests.R and you ran `get_example_data()` which generates a folder called Osiris_Data_Test in your directory. Then you ran` calculate_deltas_from_climate()`? That's strange since this function shouldn't need `tas_Amon_CanESM5_historical_r1i1p1f1_orig.nc`. Let me know if you have further questions please.
