### 2023-03-24 #5226 @jsun
I'm following the vignette https://jgcri.github.io/osiris/articles/vignette.html to run the following scripts.

```
library(osiris)
get_example_data()

data_folder <- 'Osiris_Data'

# Step 1: calculate_deltas_from_climate
osiris::calculate_deltas_from_climate(
  climate_dir = paste0(data_folder,"/climate_data"),
  write_dir = paste0(data_folder,"/outputs_calculate_delta_from_climate"),
  monthly_growing_season = paste0(data_folder,"/growing_seasons/pmonth_gslength_unifWheat_smallareamask.csv"),
  monthly_harvest_season = paste0(data_folder,"/growing_seasons/p_h_months_unifWheat_smallareamask.csv"),
  growing_season_dir = paste0(data_folder,"/growing_season_climate_data"),
  esm_name = "CanESM5",
  crops = c("Corn", "Wheat", "Rice", "Soy"),
  irrigation_rainfed = c("IRR", "RFD"),
  minlat = -87.8638,
  minlon = -179.75,
  rollingAvgYears = 15,
  tas_historical = "tas_Amon_CanESM5_historical_r1i1p1f1_orig.nc",
  tas_projected = "tas_Amon_CanESM5_ssp585_r1i1p1f1_gn_201501-210012.nc",
  pr_historical = "pr_Amon_CanESM5_historical_r1i1p1f1_orig.nc",
  pr_projected = "pr_Amon_CanESM5_ssp585_r1i1p1f1_gn_201501-210012.nc",
  historical_start_year = 1850,
  projection_start_year = 2015
)
```

The version of osiris is osiris_1.0.0 running on "R version 4.2.0 (2022-04-22), Platform: x86_64-apple-darwin17.0 (64-bit), Running under macOS Big Sur/Monterey 10.16".
