+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 2.0.7
 pkgdown_sha: ~
 articles:
   vignette: vignette.html
-last_built: 2023-02-24T23:20Z
+last_built: 2023-03-24T16:36Z
 
