### 2023-03-27 #5226 @jsun
Hi, I updated the package and followed the online vignette to try to run the package. I still got the following error (it seems that the other file is not found).

```
Starting calculate_deltas_from_climate...
Error in R_nc4_open: No such file or directory
Error in ncdf4::nc_open(batch0[1]) :
  Error in nc_open trying to open file Osiris_Data/climate_data/tas_Amon_CanESM5_historical_r1i1p1f1_gn_201001-201501.nc (return_on_error= FALSE )
In addition: Warning message:
'memory.limit()' is Windows-specific
```
