### 2023-03-27 a4d603c
Updated get_example_data function so that it installs Osiris_Data_test
