+++ DESCRIPTION
@@ -14,7 +14,7 @@ Depends: R (>= 2.10)
 License: BSD_2_clause + file LICENSE
 Encoding: UTF-8
 LazyData: true
-RoxygenNote: 7.2.1
+RoxygenNote: 7.2.3
 Imports: 
 	dplyr,
 	data.table,
