### 2023-03-27 #5226 @hahsan1
The issue is fixed now. The `get_example_data()` function will install a folder called Osiris_Data_Test (which is smaller).

After running that function, just set data_folder <- 'Osiris_Data_Test', then you can continue.

The outputs for each step are generated in the Osiris_Data_Test folder and they will be called outputs_<name_of_function>. You will also see folders called outputs_<name_of_function>_test. These come with the data set and are used by the osiris/tests/testthat/test-osiris.R script for doing checks for ensuring accuracy of results.
