### 2023-03-29 #5226 @jsun
Hi, thanks to fix the downloading path. I could successfully execute all functions following the vignette. Here is my question about the osiris package. The use of this package is described in detail in the vignette and there are four core functions for executing GCAM. However, each function requires a number of files for input; and reading the function documentation did not help me to understand the format of these files. For example,

- The `calculate_deltas_from_climate` function requires a `climate_dir`, but there is no description of what files should be in the `climate_dir`.
- The `monthly_growing_season` requires a csv file with columns named latgrid, ..., areamask; but I can't find definitions of these column names (e.g., what is `areamsk').

The complex inputs and outputs of these functions in the osiris package gave me the impression that this package is intended to solve a specific problem and that other researchers cannot use it to solve their own problems.  So my question is does the package intend to solve a specific problem (e.g., the input data must be downloaded from xxx, crop must be xxx, ... etc)? If so, it should be mentioned that the osiris package is intended to solve a specific problem in the manuscript. Whereas, if the package is intended to be used by others, it is advisable to write a more detailed description of how to prepare the input files.
