### 2023-04-03 #5226 @hahsan1
Thank you for the questions.
I can add more detail in the user guide and a description column in the tables to offer more details about the file contents.

Some of the input files are flexible in terms of the user being able to exchange the default file with their own file, and I think we can make that clearer. For example, the `climate_dir` contains the climate data that the user wants to apply to the ag yield. So it needs a historical and projection netcdf each for temperature (K) and precipitation (kg m-2 s-1) on a global map. In our current test dataset, we are using data from the CanESM5 model (historical and ssp245/ssp585 projections).These files can be easily accessed from CMIP6 for download via either ESGF (https://esgf-node.llnl.gov/projects/cmip6/) or Pangeo (https://gallery.pangeo.io/repos/pangeo-gallery/cmip6/intake_ESM_example.html). We will update the description

The crop yield emulators in `yield_response_fcns` are also updateable. However, the package is currently geared towards the AgMIP GGCMI Phase 2 style of emulators, with the emulation of the gridded crop model LPJmL included as the default. This emulator has a yield response function for corn, rice, soy and wheat.

The files in the `growing_seasons` folder were specifically used for a previous study, and they are the default files we use. They can however be updated for other growing season assumptions

The files in `gcamdata_files`, `mapping_data`, and `reference_agprodchange` are essentially specific to the version of GCAM being used and do not change once the gcamdata system for that version has been built.

The specific problem this package solves is getting gridded yield information into the GCAM style files in a way that is efficient and reproducible. If the user does not have gridded yield information, then the emulator can produce that using climate data from cmip (or isimip). Response functions are generally available in the literature. Best practices would be to have responses for at least Corn, Wheat, Rice, and Soy although the commodity mapping the package takes can be updated to include other crops.
