### 2023-04-05 3a66601
Added details to Table 1 in vignette
