+++ vignettes/vignette.Rmd
@@ -118,22 +118,28 @@ devtools::install_github("JGCRI/osiris")
 
 Additional steps for UBUNTU from a terminal
 
-    sudo add-apt-repository ppa:ubuntugis/ppa
-    sudo apt-get update
-    sudo apt-get install -y libcurl4-openssl-dev libssl-dev libxml2-dev libudunits2-dev libavfilter-dev  libmagick++-dev
+```         
+sudo add-apt-repository ppa:ubuntugis/ppa
+sudo apt-get update
+sudo apt-get install -y libcurl4-openssl-dev libssl-dev libxml2-dev libudunits2-dev libavfilter-dev  libmagick++-dev
+```
 
 Additional steps for MACOSX from a terminal
 
-    brew install pkg-config
-    
+```         
+brew install pkg-config
+```
+
 ### Download supplement
+
 To run the Osiris package, download the supplementary data which includes the climate data files, mapping files, crop emulator files, etc. Note that the `climate_data` folder currently contains temperature and precipitation files specifically from the CanESM5 model. The Osiris package was originally designed to function with the CanESM5 data. If a different set of climate data are used, it may need to be regridded to conform to the same format as CanESM5. Feel free to start an issue on the GitHub page or contact the author (Hamza) for help needed to integrate your climate data into Osiris.
 
 ``` r
 library(osiris)
 get_example_data()
 ```
-The `get_example_data()` function takes in two arguments, `write_dir` (default is the current directory) and `data_link` (default is the zenodo link to the Osiris data). 
+
+The `get_example_data()` function takes in two arguments, `write_dir` (default is the current directory) and `data_link` (default is the zenodo link to the Osiris data).
 
 <!-------------------------->
 
@@ -169,13 +175,55 @@ The first step in the workflow processes climate data (see table for source) for
 
 ## Table
 
+<table style="width:100%">
+<thead>
+  <tr>
+    <th>File</th>
+    <th style="width:50%">Description</th>
+    <th>Reference</th>
+    <th>Source</th>
+  </tr>
+</thead>
+<tbody>
+  <tr>
+    <td>pr_Amon_CanESM5_historical_r1i1p1f1.nc <br>tas_Amon_CanESM5_historical_r1i1p1f1.nc</td>
+    <td rowspan="3">Netcdf of historical and future projected global temperature (K) and precipitation (kg m-2 s-1).</td>
+    <td rowspan="3">Swart et al., 2019</td>
+    <td>https://www.wdc-climate.de/ui/cmip6?input=CMIP6.CMIP.CCCma.CanESM5.historical</td>
+  </tr>
+  <tr>
+    <td>pr_Amon_CanESM5_ssp245_r1i1p1f1.nc <br>tas_Amon_CanESM5_ssp245_r1i1p1f1.nc</td>
+    <td>https://www.wdc-climate.de/ui/cmip6?input=CMIP6.ScenarioMIP.CCCma.CanESM5.ssp245</td>
+  </tr>
+  <tr>
+    <td>pr_Amon_CanESM5_ssp585_r1i1p1f1.nc <br>tas_Amon_CanESM5_ssp585_r1i1p1f1.nc</td>
+    <td>https://www.wdc-climate.de/ui/cmip6?input=CMIP6.ScenarioMIP.CCCma.CanESM5.ssp585</td>
+  </tr>
+  <tr>
+    <td>pmonth_gslength_unifWheat_smallareamask.csv</td>
+    <td rowspan="2">Data for planting and harvesting months, and growing season length. Default files taken from previous study. These can be updated for other growing season assumptions.</td>
+    <td rowspan="2">Snyder, 2018</td>
[…]
