### 2023-04-10 JGCRI/osiris#24 @kauedesousa
get_example_data() always download
This issue belongs to the review on JOSS https://github.com/openjournals/joss-reviews/issues/5226

The function `get_example_data()` always download the data even when it already exists in the directory. My suggestion is to dig into the function `geodata::gadm()` and check their approach to download the data only once even if you run the line of code many times. Additionally is nice to return a message highlighting the path where the folder was written.

```
library("geodata")

gadm("ZWE", level=0, path = getwd())

```
