### 2023-04-10 JGCRI/osiris#25 @kauedesousa
Avoid pipes %>% in functions
This issue belongs to the review on JOSS openjournals/joss-reviews#5226

Pipes %>% are nice when writing scripts. For R packages is not. They are difficult to control inside a function and also increases the computational cost. Even the `tidyverse` friend package `sf` don't use it in their code https://github.com/r-spatial/sf/blob/main/R/tidyverse.R. Please remove the pipes.
