### 2023-04-10 JGCRI/osiris#25 @kauedesousa
And it becomes really strange (and dangerous) when you assign all the initial objects as `.`

Like here   [NULL -> year -> value -> vtag -> units -> ctag -> type -> Scenario ->](https://github.com/JGCRI/osiris/blob/a54410f4a6a685585ec084cf0a42cf969d42357f/R/grid_to_basin_yield.R#L58)
