### 2023-04-21 49f3e20
Added error message if user inputs incorrect crop
