### 2023-04-21 JGCRI/osiris#26 @hahsan1
Thank you for this suggestion. I have incorporated an error message into these two functions. Now, if the user inputs any crop that is not listed as one of the default crops for the respective functions, the script will abort and the following error message will show up:

Error:  The crops that were defined are either misspelled or not included in this package.
         The crops currently included are: Corn, Wheat, Rice, Soy
