### 2023-04-23 JGCRI/osiris#24 @hahsan1
Thank you for this suggestion. I have updated the `get_example_data()` function so that it now outputs a message to let the user know if the sample data is already installed. I also made it clear which path the data is installed to.
