### 2023-04-23 7ade201
Added message to let user know if sample data already exists when attempting to rerun get_example_data function.
