+++ R/get_example_data.R
@@ -27,29 +27,35 @@ get_example_data <- function(write_dir = getwd(),
 
   rlang::inform("Starting get_example_data")
 
-  # Download zip file from zenodo
-  rlang::inform("Starting download...")
+  # Check if sample data already exists in current directory
+  if (!dir.exists(paths = paste0(getwd(), "/Osiris_Data_Test"))) {
+    # Download zip file from zenodo
+    rlang::inform("Starting download...")
 
-  options(timeout = max(300, getOption("timeout")))
+    options(timeout = max(300, getOption("timeout")))
 
-  utils::download.file(url = data_link,
-                       destfile = paste0(write_dir, "/", dir_name, ".zip"),
-                       mode = "wb")
+    utils::download.file(url = data_link,
+                         destfile = paste0(write_dir, "/", dir_name, ".zip"),
+                         mode = "wb")
 
-  rlang::inform("Download complete.")
+    rlang::inform("Download complete.")
 
-  # Download zip file from zenodo
-  rlang::inform("Starting unzip...")
+    # Download zip file from zenodo
+    rlang::inform("Starting unzip...")
 
-  utils::unzip(paste0(write_dir, "/", dir_name, ".zip"),
-               exdir = paste0(write_dir, "/", dir_name))
+    utils::unzip(paste0(write_dir, "/", dir_name, ".zip"),
+                 exdir = paste0(write_dir, "/", dir_name))
 
-  unlink(paste0(write_dir, "/", dir_name, ".zip"))
+    unlink(paste0(write_dir, "/", dir_name, ".zip"))
 
-  rlang::inform("Unzip complete.")
+    rlang::inform("Unzip complete.")
 
-  rlang::inform(paste0("You can set data_folder = \"", write_dir, "/", dir_name, "\" to run the example scripts from https://jgcri.github.io/osiris/articles/vignette.html"))
-
+    rlang::inform(paste0("The example data was installed to: ", getwd(), "/Osiris_Data_Test
+You can set data_folder = \"", write_dir, "/", dir_name, "\" to run the example scripts from https://jgcri.github.io/osiris/articles/vignette.html"))
+  } else {
+    rlang::inform(paste0("Example data folder already exists.
+If you would like to reinstall the data, delete the ", getwd(), "/Osiris_Data_Test folder and run this function again."))
+  }
 
   #.........................
   # Close Out
