### 2023-04-24 JGCRI/osiris#25 @hahsan1
> This issue belongs to the review on JOSS [openjournals/joss-reviews#5226](https://github.com/openjournals/joss-reviews/issues/5226)
>
> Pipes %>% are nice when writing scripts. For R packages is not. They are difficult to control inside a function and also increases the computational cost. Even the `tidyverse` friend package `sf` don't use it in their code https://github.com/r-spatial/sf/blob/main/R/tidyverse.R. Please remove the pipes.

We appreciate your comments regarding the use of pipes in our functions. You bring up a good point about increase in computational cost. However, we believe that the use of pipes greatly improves the readability and flow of the code. Without the pipes, the code would be overly verbose and difficult to follow. We believe keeping the code readable and accessible to humans is an important factor to consider besides the pure computational aspects.
