### 2023-05-12 22ca333
Fixed typo in paper and edited figure 1
