+++ paper/paper.md
@@ -34,7 +34,7 @@ bibliography: paper.bib
 
 # Summary
 
-`osiris` is an R package that couples climate change projections with socioeconomic agricultural assumptions used in the Global Change Analysis Model (GCAM). Specifically, this package allows users to examine the implications of changes to regional or global temperature and precipitation for agricultural crop production using GCAM. `osiris` is designed for modularity to allow the user to explore alternative climate projections, crop model emulators, and parameters relating to agricultural yield. The development of this package is part of a broader effort to to provide reproducible, updatable data processing for analyzing the impacts of climate [@Jones2023] and socioeconomic change on the coevolution of future water, energy, and land systems using GCAM. Figure 1 shows a generic workflow where climate data are processed by `osiris` and a set of agricultural productivity change files are generated as inputs for GCAM.
+`osiris` is an R package that couples climate change projections with socioeconomic agricultural assumptions used in the Global Change Analysis Model (GCAM). Specifically, this package allows users to examine the implications of changes to regional or global temperature and precipitation for agricultural crop production using GCAM. `osiris` is designed for modularity to allow the user to explore alternative climate projections, crop model emulators, and parameters relating to agricultural yield. The development of this package is part of a broader effort to provide reproducible, updatable data processing for analyzing the impacts of climate [@Jones2023] and socioeconomic change on the coevolution of future water, energy, and land systems using GCAM. Figure 1 shows a generic workflow where climate data are processed by `osiris` and a set of agricultural productivity change files are generated as inputs for GCAM.
 
 ![Climate data are processed to generate agricultural productivity change inputs for GCAM.](osiris_workflow.png)
 
@@ -42,7 +42,7 @@ bibliography: paper.bib
 
 Changes in temperature, precipitation, length of growing seasons, and *CO~2~* concentration have been shown to have a potential impact on agricultural crop yields in the future [@Asseng2015]. Studies suggest that although the projections of future yields are uncertain, they may affect major agricultural producers in terms of crop production quantity and composition, as well as global trade and market value [@Snyder2020; @Baker2018]. Central to the analysis of the impact of climate on agriculture are tools that can efficiently process large data sets from various sources and generate reproducible results that can be used within a modeling framework.
 
-`osiris` is an R package developed with the aim to process climate impacts on agricultural yields for the GCAM model [@Calvin2019]. This package is a modular data pipeline that can take climate data at high spatiotemporal resolution (e.g., hourly meteorological variables at 12 km resolution over continental United States (CONUS)) and gridded crop yield data from existing emulators [@Franke2020], and transforms the gridded yield time series into yield impacts for the agricultural commodities modeled in GCAM. From a set of gridded crop yield time series (e.g., from the emulator stage), the data processing steps to GCAM input files are relatively agnostic of the source of the gridded yield time series. The gridded yield time series to GCAM input file portion of the data processing pipeline in `osiris` follows the approach taken in past studies examining agricultural impacts in GCAM [@Kyle2014; @Snyder2020; @Graham2020; @Zhao2021]. The inclusion of these operations in the `osiris` package as functions rather than static scripts that must be updated by project or hardcoded for processing of specific yield scenarios in the gcamdata system [@Lamberty2019], is an obvious improvement in the adherence of GCAM studies to FAIR data principles (Findable, Accessible, Interoperable, and Reusable), in addition to offering flexibility for future updates to be integrated into GCAM runs more easily.
[…]
