+++ paper/paper.bib
@@ -15,7 +15,7 @@
 }
 
 @article{Baker2018,
-	title = {Evaluating the effects of climate change on US agricultural systems: sensitivity to regional impact and trade expansion scenarios},
+	title = {Evaluating the effects of climate change on {US} agricultural systems: {sensitivity} to regional impact and trade expansion scenarios},
 	volume = {13},
 	url = {https://iopscience.iop.org/article/10.1088/1748-9326/aac1c2},
 	doi = {10.1088/1748-9326/aac1c2},
@@ -30,7 +30,7 @@
 }
 
 @article{Snyder2020,
-	title = {The domestic and international implications of future climate for U.S. agriculture in GCAM},
+	title = {The domestic and international implications of future climate for {U.S.} agriculture in {GCAM}},
 	volume = {15},
 	url = {https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0237918},
 	doi = {10.1371/journal.pone.0237918},
@@ -54,12 +54,12 @@
   number = {2},
   pages = {677--698},
   author = {Katherine Calvin and Pralit Patel and Leon Clarke and Ghassem Asrar and Ben Bond-Lamberty and Ryna Yiyun Cui and Alan Di Vittorio and Kalyn Dorheim and Jae Edmonds and Corinne Hartin and Mohamad Hejazi and Russell Horowitz and Gokul Iyer and Page Kyle and Sonny Kim and Robert Link and Haewon McJeon and Steven J. Smith and Abigail Snyder and Stephanie Waldhoff and Marshall Wise},
-  title = {{GCAM} v5.1: representing the linkages between energy,  water,  land,  climate,  and economic systems},
+  title = {{GCAM} v5.1: {representing} the linkages between energy, water, land, climate, and economic systems},
   journal = {Geoscientific Model Development}
 }
 
 @article{Franke2020,
-	title = {The GGCMI Phase 2 emulators: global gridded crop model responses to changes in CO2, temperature, water, and nitrogen (version 1.0)},
+	title = {The {GGCMI} {Phase} 2 emulators: {global} gridded crop model responses to changes in {CO$_2$}, temperature, water, and nitrogen (version 1.0)},
 	volume = {13},
 	url = {https://gmd.copernicus.org/articles/13/3995/2020/},
 	doi = {10.5194/gmd-13-3995-2020},
@@ -74,7 +74,7 @@
 }
 
 @article{Portmann2010,
-	title = {MIRCA2000—Global monthly irrigated and rainfed crop areas around the year 2000: A new high-resolution data set for agricultural and hydrological modeling},
+	title = {{MIRCA2000}—{Global} monthly irrigated and rainfed crop areas around the year 2000: {A} new high-resolution data set for agricultural and hydrological modeling},
 	volume = {24},
 	url = {https://agupubs.onlinelibrary.wiley.com/doi/full/10.1029/2008GB003435},
 	doi = {10.1029/2008GB003435},
@@ -88,7 +88,7 @@
 }
 
 @article{Lamberty2019,
-	title = {gcamdata: An R Package for Preparation, Synthesis, and Tracking of Input Data for the GCAM Integrated Human-Earth Systems Model},
+	title = {\emph{gcamdata}: {An} {R} {Package} for {Preparation}, {Synthesis}, and {Tracking} of {Input} {Data} for the {GCAM} {Integrated} {Human-Earth} {Systems} {Model}},
 	volume = {7},
 	url = {https://openresearchsoftware.metajnl.com/article/10.5334/jors.232/},
 	doi = {10.5334/jors.232},
@@ -102,7 +102,7 @@
 }
 
 @article{Jones2023,
-	title = {Continental United States climate projections based on thermodynamic modification of historical weather},
[…]
