### 2023-05-19 #5226 @hahsan1
I've made changes to the .bib file as suggested. The changes don't seem to appear on the article proof I generated though.
