### 2023-05-19 c0e1d98
Updated reference bib file
