+++ paper/paper.bib
@@ -30,7 +30,7 @@
 }
 
 @article{Snyder2020,
-	title = {The domestic and international implications of future climate for {U.S.} agriculture in {GCAM}},
+	title = {The domestic and international implications of future climate for {U.S.} {agriculture} in {GCAM}},
 	volume = {15},
 	url = {https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0237918},
 	doi = {10.1371/journal.pone.0237918},
@@ -88,7 +88,7 @@
 }
 
 @article{Lamberty2019,
-	title = {\emph{gcamdata}: {An} {R} {Package} for {Preparation}, {Synthesis}, and {Tracking} of {Input} {Data} for the {GCAM} {Integrated} {Human-Earth} {Systems} {Model}},
+	title = {{\emph{gcamdata}}: {An} {R} {Package} for {Preparation}, {Synthesis}, and {Tracking} of {Input} {Data} for the {GCAM} {Integrated} {Human-Earth} {Systems} {Model}},
 	volume = {7},
 	url = {https://openresearchsoftware.metajnl.com/article/10.5334/jors.232/},
 	doi = {10.5334/jors.232},
